### 2020-07-17 #45 @IainDaviesMaths
[Re] The principal components of natural images
**Original article:** The Principal Components of Natural Images by Hancock et al (1991),
http://citeseerx.ist.psu.edu/viewdoc/download;jsessionid=044DFF560FC2E801B579C2F23D268B44?doi=10.1.1.41.192&rep=rep1&type=pdf*

**Scientific domain:** Computational Neuroscience
**Programming language:** Julia
**Suggested editor:** -
