### 2020-07-17 #45 @cJarvers
Oh, this looks interesting. @(whoever becomes editor for this), I'd be happy to review.
