### 2020-09-08 #45 @pberkes
The authors successfully reproduce the findings of Hancock’s  “PCA of natural images” paper, with minor deviations from the original results that are easily accounted for. The paper is well written and well organized. It’s good work, I don’t have any further comment.

The code is provided in two languages, Matlab and Julia. Unfortunately, I do not have a Matlab license so I could not verify that that part of the code works as intended. As far as I’m concerned, the Matlab code could be removed from the repository as it is redundant and Matlab is not an open language.

The Julia code is readable and decently organized. It runs as intended and produces *some version* of the figure in the paper. The random seed has not been fixed, so it produces slightly different images for each run.

For instance, this is the image shown in the paper and shipped with the repository:
![julia-fig6 copy](https://user-images.githubusercontent.com/314633/92486541-838f3d80-f1ec-11ea-83bf-d0dd5b8d280e.png)

Running the code I obtain this version
￼![julia-fig6](https://user-images.githubusercontent.com/314633/92486560-8853f180-f1ec-11ea-84ef-34a53424b12c.png)

I don’t know where the editors stand on this issue, but personally I think that the random seed should be a global parameter, and running the code with the same setting would always return exactly the same images.

Regarding the code repository: while I do appreciate the effort of the authors in making their research reproducible, the GitHub repository could be organized better.
- Use directories to organize the content: input data, output results, Matlab, and Julia code in separate directories
- Please add some instructions for running the code. E.g. for Julia one needs to run `setup.jl` first and then  `make run-julia`
- The version of packages that need to be installed for Julia is specified in Project.toml, it would be useful to also have the Matlab version and a list of any Matlab optional packages

Thank you for reproducing this classic!
