### 2020-09-08 #45 @otizonaizit
@IainDaviesMaths :
  -  I agree with @pberkes that the random seed should be a global parameter
  - Regarding the Matlab code, I'd like to wait for @cJarvers and see if they can review it. In that case the code should stay. If not, then we may need to find someone who has a Matlab license and can review the code...
  - Can you comment/act on the other observations by @pberkes ?
