### 2020-09-10 #45 @cJarvers
@otizonaizit I'll try to do it within the next few days. Since I have a 6-month-old at home, it's a bit hard to predict how much time I can actually spend on work each day, but I'll do it by end of next week the latest.
I have a Matlab license, so I should be able to run & review that part of the code as well.
