### 2020-09-16 #45 @cJarvers
# Review of Davies & Eglen: [Re] The principal components of natural images

Here is my review, @otizonaizit  & @IainDaviesMaths

This review based on commit https://github.com/IainDaviesMaths/Reproduction-Hancock/commit/519b1695eb05a1d328b033728025c5e90f1d40c5

## Success of the replication

The authors successfully replicated the results of the original article. There are some minor deviations, but these are within the range of what should be expected, given that the additional natural images and the text images were not available anymore. The authors discuss these differences sufficiently.

The only place where I feel that more detailed commentary would be helpful is the reproduction of the text experiment. The authors state that "[f]or this to be reproduced, further experimental modifications had to be made to the text image", but what exactly these modifications were remains unclear. A suitable font had to be chosen (which one?) - anything else? Since the authors state that these results seem to be brittle, it might be nice to also include the result for a different font.

## Reproducibility

The authors include a makefile, which makes rerunning their code easy. Since the code is provided in both Matlab and Julia, I'll briefly comment on each below.

**Julia:**

I ran the code with Julia 1.5.1.

As @pberkes already noted, the Julia code runs without problems, but does not produce perfectly reproducible results. I agree that the most crucial fix would be to set the random seeds.

Another potential problem for reproducibility is package management. The authors provide a `setup.jl` script that installs the required dependencies. However, the versions that get installed are not necessarily the same ones as the authors used, so the behavior of the code might change in the future. However, the repo already contains the solution to this problem: the `Manifest.toml` and `Project.toml` files document the exact package versions that were used. Instead of letting the user run the `setup.jl` script, the authors could add the following lines to the beginning of their `runall.jl` script:

```
using Pkg
Pkg.activate(".")
Pkg.instantiate()
```

In that case it might also be good to remove the `setup.jl` script to avoid confusion.

**Matlab:**

I ran the code with Matlab version R2019b.

Similar to the Julia code, the Matlab version runs without problems and generates Figures that resemble those in the paper, but have subtle differences (for example, the sign of he principal components is flipped) since the random seeds are not fixed. As with the Julia code, this should be rectified by setting random seeds.

Running `make all` generates figures 2 to 8 with Matlab, leaving out figure 1. Since figure 1 only concatenates the input images, this is not a big problem, but for completeness sake it might be better to generate this figure as well (just like the Julia code).

Since Matlab is not an open language, I also tried running the code with Octave 5.2.0. This did not throw any errors and the results for figure 1 were correct, but the other figures ran for a long time. I gave up after a few hours. I know it should be possible to make it run faster in Octave using the JIT compiler or something, but I don't have enough experience with that to go down that road.

As far as I understand it, the policy of ReScience is that the code has to be in an open source language, so not Matlab. This replication is an interesting edge case, since there is also an open source implementation (in Julia). The question is whether both should be published. I guess this is for the editorial team to decide.

## Clarity of the code

I agree with @pberkes that the structure of the repository should be cleaned up. There should be separate subfolders for the images, the generated figures, the Julia code and the Matlab code.

In general, the code is clear and easy to read. There are only a few improvements I would suggest. All of these are suggestions, not requirements. If the authors disagree with me, I would nevertheless recommend publication.

**Matlab:**

In the Matlab code, there are some code duplications. For example, the functions `LearningProcess` and `LearningProcessRotated` are nearly identical, except for the rotation in the latter. One could just make rotation a parameter of the function and pass `0` if the images should not be rotated. Another example occurs in `PlotFigure7.m` and `PlotFigure8.m`. Here, the calculation of the mean gray level and the learning process are included in the script, even though the code is basically identical to the code in the functions `MeanGrayLevel` and `LearningProcess`. It would be better to just replace the code blocks by calls to these functions.

In general, I think the Matlab code would be a bit more readable if spaces were added around operators.

**Julia:**

Since the Julia code is very close to the Matlab code, the same improvements (reduction of code duplications, more spaces) could be made here. One additional recommendation concerns the documentation. Currently, each function is documented with an initial comment, Matlab-style. The convention in Julia is to use docstrings [(see here)](https://docs.julialang.org/en/v1/manual/documentation/). This has the advantage that Julias inbuilt help features (e.g., at the REPL) can be used. So simply cutting the text of each documentation comment and pasting it as a string atop the function would be an improvement. Again, this is only a recommendation.

## Clarity of the article

The article is excellent. It is easy to understand what was done, which parts of the replication were difficult etc.

## Some small issues

While reading the paper and code, I found a few small things (typos etc.) that could be improved:

- In the introduction, the original paper is cited as "Hancock et al. 1991", but in the references the date is given as 1992.
- In figure 7, the example image at half scale seems to be centered on a space, so that no letters are visible. It would be better to actually show letters here. Also the figure caption references (a), (b) and (c), but these letters are not actually used in the figure.
- The file `sagerupdate.m` should probably be called `sangerupdate.m`
- In that file in the comment on line 16, `Sager` should be `Sanger`
- In `PlotFigure6.m` the comment in line 36 says "Double scale", but should probably be "Half scale"
- The same holds for `hancock.jl` line 469
- The code in lines 378, 379, 798-800, 803, 817-820 and 835-837 in `hancock.jl` has been commented out and could be removed to increase clarity

## TL;DR

The replication was successful. The article is excellent. The code is also well structured and easily readable, however there are two problems with reproducibility (lack of random seeds; suboptimal package management) that the authors should fix before publication. In addition, the repo should be cleaned up.

Since one of the two implementation is in Matlab, the editorial team needs to decide whether that is appropriate for publication in ReScience, or whether only the Julia code should be published.

Finally, there are some improvements to the code that I would recommend (reducing duplications, adding spaces around operators, adding docstrings), but these are only suggestions.
