### 2020-10-04 b14d848
Re-organised repo and made suggested changes to MATLAB code. Also added program to create PCs for text with different font
