### 2020-10-04 5a9c3a7
Made changes to Julia code to remove LearningProcessRotated and duplicate code in PlotFigure7 and 8. Also added PlotFigure9
