+++ julia-code/hancock.jl
@@ -1,12 +1,12 @@
 function SaveStandardImageBase()
     #Save the 15 standard images into standard_image_base.
-    standard_image_base=zeros(256,256,15);
+    standard_image_base = zeros(256,256,15);
 
     for p = 1:15
-        file = @sprintf("images%d.gif", p)
-        image=load(file)
+        file = @sprintf("input-images/images%d.gif", p)
+        image = load(file)
         i = Gray.(image)
-        standard_image_base[:,:,p]=i
+        standard_image_base[:,:,p] = i
     end
 
     return standard_image_base
@@ -15,8 +15,8 @@ end
 
 function PlotFigure1()
     standard_image_base = SaveStandardImageBase()
-    mosaic = mosaicview(standard_image_base; nrow=3)
-    save("julia-fig1.png", mosaic)
+    mosaic = mosaicview(standard_image_base; nrow = 3)
+    save("julia-figs/julia-fig1.png", mosaic)
 end
 
 
@@ -26,16 +26,19 @@ function MeanGreyLevel(side_pixels, image_base, num_image)
     ## MEANGREYLEVEL finds mean grey level of 20000 samples of side_pixel^2
     ## samples of images.
 
-    
-    greylevel=0.0;
-    for I=1:20000
+	# Extract dimensions of images in the image_base. All images in the 
+	# image_base must have same dimension.
+	(dim_x, dim_y) = size(image_base);
+	    
+    greylevel = 0.0;
+    for I = 1:20000
         # Select piece of image from database.
-        image_no=rand(1:num_image);
-        image_x=rand(1:(257-side_pixels));
-        image_y=rand(1:(257-side_pixels));
-        image=image_base[image_x:image_x+side_pixels-1,image_y:image_y+side_pixels-1,image_no];
-        # Find average grey level.
-        greylevel=greylevel+sum(sum(image/(side_pixels^2)))/20000;
+        image_no = rand(1:num_image);
+		image_x = rand(1:dim_x+1-side_pixels);
+		image_y = rand(1:dim_y+1-side_pixels);    
+		image = image_base[image_x:image_x+side_pixels-1,image_y:image_y+side_pixels-1,image_no];
+		# Find average grey level.
+        greylevel = greylevel+sum(sum(image/(side_pixels^2)))/20000;
     end
     return greylevel
 end
@@ -43,7 +46,7 @@ end
 
 
 
-function LearningProcess(num_unit, side_pixels, image_base, num_image,meangrey)
+function LearningProcess(num_unit, side_pixels, image_base, num_image,meangrey, rotation_angle)
     ##LEARNINGPROCESS simulates the learning process of the network according to
     ##Sanger's rule
 
@@ -53,32 +56,40 @@ function LearningProcess(num_unit, side_pixels, image_base, num_image,meangrey)
     ##num_image      Number of images in the database.
     ##meangrey       The mean grey level of 20000 samples of the given size from
     ##               the database. 
+    ##rotation_angle Images will be rotated by this number of degrees
+
+	##Extract dimensions of images in the image_base. All images in the 
+	##image_base must have same dimension.
+	(dim_x, dim_y) = size(image_base);
 
     ##Initialise weights so that the sum of the squares is approximately 1.
-    weights=0.06*(rand(num_unit,side_pixels^2) .- 0.5);
+    weights = 0.06*(rand(num_unit,side_pixels^2) .- 0.5);
     
-    learn_rate=1;
-    for time=1:120000
+    learn_rate = 1;
+    for time = 1:120000
         ##Select 64x64 piece of image from database.
-        image_no=rand(1:num_image);
-        image_x=rand(1:(257-side_pixels));
-        image_y=rand(1:(257-side_pixels));
[…]
