### 2020-10-04 0e75e24
Added Figure 9 to Makefile and made some edits to the code to produce Figure 9
