### 2020-10-04 #45 @IainDaviesMaths
Thank you everyone for the reviews and feedback on the paper! Myself and Stephen are in the process of making all the suggested edits and should be done in the next few days.
