### 2020-10-05 3644b5c
Added MATLAB dependencies to README and explained how to use Julia code
