### 2020-10-05 4615d66
made minor change to Julia code to get correct contrast on PCs
