### 2020-10-05 #45 @IainDaviesMaths
Most edits have been done now - just a few minor Julia edits to be made which Stephen is on top of. Will alert when the article is in final draft.
