### 2020-10-21 56a0431
Add docstrings and save outputs to current directory.
