+++ julia-code/hancock.jl
@@ -1,9 +1,15 @@
+"Seed the RNG to a fixed value.  For reproducibility."
+function seedRNG()
+    Random.seed!(1234);    
+end
+
+
+"Save the 15 standard images into standard_image_base."
 function SaveStandardImageBase()
-    #Save the 15 standard images into standard_image_base.
     standard_image_base = zeros(256,256,15);
 
     for p = 1:15
-        file = @sprintf("input-images/images%d.gif", p)
+        file = @sprintf("../input-images/images%d.gif", p)
         image = load(file)
         i = Gray.(image)
         standard_image_base[:,:,p] = i
@@ -13,6 +19,7 @@ function SaveStandardImageBase()
 end
 
 
+"Plot Figure 1"
 function PlotFigure1()
     standard_image_base = SaveStandardImageBase()
     mosaic = mosaicview(standard_image_base; nrow = 3)
@@ -21,11 +28,8 @@ end
 
 
 
-######################################################################
+"Find mean grey level of 20000 samples of side_pixel^2 samples of images."
 function MeanGreyLevel(side_pixels, image_base, num_image)
-    ## MEANGREYLEVEL finds mean grey level of 20000 samples of side_pixel^2
-    ## samples of images.
-
 	# Extract dimensions of images in the image_base. All images in the 
 	# image_base must have same dimension.
 	(dim_x, dim_y) = size(image_base);
@@ -45,18 +49,18 @@ end
 
 
 
+"""
+Simulate the learning process of the network according to Sanger's rule.
 
+num_unit       Number of output units in the network.
+side_pixels    Number of pixels in a side of the presented images.
+image_base     The database of images to be sampled from.
+num_image      Number of images in the database.
+meangrey       The mean grey level of 20000 samples of the given size from
+               the database. 
+rotation_angle Images will be rotated by this number of degrees
+"""
 function LearningProcess(num_unit, side_pixels, image_base, num_image,meangrey, rotation_angle)
-    ##LEARNINGPROCESS simulates the learning process of the network according to
-    ##Sanger's rule
-
-    ##num_unit       Number of output units in the network.
-    ##side_pixels    Number of pixels in a side of the presented images.
-    ##image_base     The database of images to be sampled from.
-    ##num_image      Number of images in the database.
-    ##meangrey       The mean grey level of 20000 samples of the given size from
-    ##               the database. 
-    ##rotation_angle Images will be rotated by this number of degrees
 
 	##Extract dimensions of images in the image_base. All images in the 
 	##image_base must have same dimension.
@@ -97,11 +101,8 @@ function LearningProcess(num_unit, side_pixels, image_base, num_image,meangrey, 
     
 end
 
-
+"Create a SIZE by SIZE array with gaussian values centred in the middle."
 function Gaussian( size )
-    ## GAUSSIAN creates a size by size array with gaussian values centred in the
-    ## middle.
-
     sigma = size*5.0/32;
 
     array = zeros(size,size);
@@ -118,11 +119,12 @@ function Gaussian( size )
 
 end
 
-
+"""
+SANGERUPDATE updates the weights of the connections between input pixels
+and output units given the presentation of an image according to Sanger's
+rule.
+"""
 function sangerupdate( image_vec, weights, learn_rate )
[…]
