### 2020-10-21 695f220
remove last two figures
