+++ julia-code/Project.toml
@@ -10,4 +10,4 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 TestImages = "5e47fb64-e119-507b-a336-dd2b206d9990"
 
 [compat]
-julia = "1.4.2"
+julia = "1.5"
