### 2020-10-21 #45 @IainDaviesMaths
Hi all - I think that's all your suggestions added to the code and article now! Let me know if there's anything else to be done.
