### 2020-10-26 #45 @cJarvers
@otizonaizit no objections from me either. The changes address all the issues I raised. Great work by the authors!
