### 2020-10-27 #45 @otizonaizit
Great! Published:

- https://github.com/ReScience/articles/pull/18
- https://github.com/ReScience/rescience.github.io/pull/95

The article will appear on https://rescience.github.io/read/ as soon as these two PRs are merged by one of the editors in chief (@rougier ?)
